merchise: Use 'safe_eval' directly.
This avoid a weird error:

Traceback (most recent call last):
  File "/home/manu/src/merchise/pgi/xhg/../openupgrade/openerp/service/server.py", line 988, in preload_registries
    registry = RegistryManager.new(dbname, update_module=update_module)
  File "/home/manu/src/merchise/pgi/xhg/../openupgrade/openerp/modules/registry.py", line 390, in new
    openerp.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/manu/src/merchise/pgi/xhg/../openupgrade/openerp/modules/loading.py", line 408, in load_modules
    force, status, report, loaded_modules, update_module, upg_registry)
  File "/home/manu/src/merchise/pgi/xhg/../openupgrade/openerp/modules/loading.py", line 297, in load_marked_modules
    graph.add_modules(cr, module_list, force)
  File "/home/manu/src/merchise/pgi/xhg/../openupgrade/openerp/modules/graph.py", line 93, in add_modules
    forced_deps = tools.safe_eval(forced_deps)
TypeError: 'module' object is not callable

What is weird about it, is that it happens after many migration steps have
been performed: so this line of code has run before.  So, it seems that
someone is changing (fixing) the value of 'tools.safe_eval' (setting the
'safe_eval' attribute of module 'tools').

I didn't trace this error.  Just avoid it.